Better hooks and simplifying
